Move from actix to axum + rspc

+++ api/src/types.rs
@@ -1,28 +1,29 @@
 use serde::{Deserialize, Serialize};
+use specta::Type;
 use sqlx::FromRow;
 
-#[derive(FromRow, Serialize)]
+#[derive(FromRow, Serialize, Type)]
 pub struct Kind {
-    pub kind_id: i64,
+    pub kind_id: i32,
     pub name: String
 }
 
-#[derive(FromRow, Serialize)]
+#[derive(FromRow, Serialize, Type)]
 pub struct Producer {
-    pub producer_id: i64,
+    pub producer_id: i32,
     pub name: String
 }
 
-#[derive(Deserialize)]
+#[derive(Deserialize, Type)]
 pub struct InsertBeverage {
     pub name: String,
-    pub producer_id: i64,
-    pub kind_id: i64,
-    pub rating: i64,
+    pub producer_id: i32,
+    pub kind_id: i32,
+    pub rating: i32,
     pub description: String
 }
 
-#[derive(FromRow, Serialize)]
+#[derive(FromRow, Serialize, Type)]
 pub struct JoinBeverage {
     pub name: String,
     pub producer: String,
